http: Move chunked data sink to common.{hh|cc}
The sink convers input flow of bytes into http chunked-encoded body. Now
it sits in the http/reply.cc, but it will be used by client code to let
user write body of unknown size into the request.
