+++ src/http/common.cc
@@ -20,6 +20,7 @@
  */
 
 #include <seastar/http/common.hh>
+#include <seastar/core/iostream-impl.hh>
 
 namespace seastar {
 
@@ -50,6 +51,55 @@ operation_type str2type(const sstring& type) {
     return GET;
 }
 
+}
+
+namespace http {
+namespace internal {
+
+class http_chunked_data_sink_impl : public data_sink_impl {
+    output_stream<char>& _out;
+
+    future<> write_size(size_t s) {
+        auto req = format("{:x}\r\n", s);
+        return _out.write(req);
+    }
+public:
+    http_chunked_data_sink_impl(output_stream<char>& out) : _out(out) {
+    }
+    virtual future<> put(net::packet data)  override { abort(); }
+    using data_sink_impl::put;
+    virtual future<> put(temporary_buffer<char> buf) override {
+        if (buf.size() == 0) {
+            // size 0 buffer should be ignored, some server
+            // may consider it an end of message
+            return make_ready_future<>();
+        }
+        auto size = buf.size();
+        return write_size(size).then([this, buf = std::move(buf)] () mutable {
+            return _out.write(buf.get(), buf.size());
+        }).then([this] () mutable {
+            return _out.write("\r\n", 2);
+        });
+    }
+    virtual future<> close() override {
+        return  make_ready_future<>();
+    }
+};
+
+class http_chunked_data_sink : public data_sink {
+public:
+    http_chunked_data_sink(output_stream<char>& out)
+        : data_sink(std::make_unique<http_chunked_data_sink_impl>(
+                out)) {}
+};
+
+output_stream<char> make_http_chunked_output_stream(output_stream<char>& out) {
+    output_stream_options opts;
+    opts.trim_to_size = true;
+    return output_stream<char>(http_chunked_data_sink(out), 32000, opts);
+}
+
+}
 }
 
 }
